refactor(bool): import new Bool instead of old Bool impl
