+++ src/lib/zkapp.ts
@@ -1,13 +1,12 @@
 import { Types } from '../bindings/mina-transaction/types.js';
 import {
-  Bool,
   Gate,
   Ledger,
   Pickles,
   Poseidon as Poseidon_,
   ProvablePure,
 } from '../snarky.js';
-import { Field } from './core.js';
+import { Field, Bool } from './core.js';
 import {
   AccountUpdate,
   AccountUpdatesLayout,
