+++ src/snarky.d.ts
@@ -1,6 +1,6 @@
 import type { Account as JsonAccount } from './bindings/mina-transaction/gen/transaction-json.js';
 import type { Field, FieldConst, FieldVar } from './lib/field.js';
-import type { BoolVar } from './lib/bool.js';
+import { type BoolVar, Bool } from './lib/bool.js';
 import type { Scalar, ScalarConst } from './lib/scalar.js';
 // export { Field };
 export { SnarkyField };
@@ -1062,8 +1062,8 @@ declare class SnarkyField {
  *
  * Use [[assertEquals]] to enforce the value of a Bool.
  */
-declare function Bool(x: Bool | boolean): Bool;
-declare class Bool {
+declare function BoolOld(x: Bool | boolean): Bool;
+declare class BoolOld {
   constructor(x: Bool | boolean);
 
   /**
